Add files via upload
This is a multi-fast file of the 1233 males in the 1000G data. Each fasta file is labeled with the ID of the sample. The nucleotides are for common variant positions in the 1000G Y chromosome VCF file with the reference allele for negative genotypes and the alternate allele for positive genotypes. The common variants are variants with AC >61, which is AF > 0.5%. Each fasta file has 1815 nucleotides. Common variants which were INDELs were excluded. Genotypes not recorded as positive or negative, are "N" in the fasta files.