Add files via upload
This multi-fasta file is made from 123 men of 1000G by randomly choosing 62 men from the 624 L_1 men, 37 men from the 373 L_2 men and 23 men from the L_3 men to make 123 men of the whole sample of 1233 men. 35 marker SNVs are used to get the nucleotide string and variants of AC above 624 are folded, that is the ancestral and alternate alleles are switched.